+++ versions.tf
@@ -1,5 +1,6 @@
 terraform {
-  required_version = ">= 0.13"
+  experiments      = [module_variable_optional_attrs]
+  required_version = ">= 0.14"
 
   required_providers {
     azurerm = {